requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pydantic[email]==2.9.2
pydantic-settings==2.5.2
pytest==8.3.3
httpx==0.27.2
pytest-cov==5.0.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

